Heal Aura rewrites.
Another attempt at making the healing and reviving less buggy.

Reduced the aura range to be more sensible.

Also altered effect of the aura pulse to give a more accurate sense of max range, as well as reducing poison damage as it was incredibly strong.

Reviving with a pulse now only costs double normal charge.